Extract HIR source location helpers
Move syntax statement/expression span accessors out of hir.rs into hir/source_locations.rs with module-private visibility.

Lower the compiler source monolith ratchet ceilings for the additional HIR facade reduction.

Refs #1254.
